requirements.txt
astroid==2.1.0
autopep8==1.4.3
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
autopep8==1.4.3
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.1.3
django-cors-middleware==1.3.1
django-extensions==2.1.4
django-heroku==0.3.1
djangorestframework==3.9.0
docopt==0.6.2
drf-yasg==1.11.1
facebook-sdk==3.1.0
future==0.17.1
get==2018.11.19
google-auth==1.6.1
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
docopt==0.6.2
get==2018.11.19
google-auth==1.6.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==2.1.0
pep8==1.7.1
Pillow==5.3.0
post==2018.11.20
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
public==2018.11.20
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.2.1
python-dotenv==0.9.1
python-twitter==3.5
pytz==2018.7
query-string==2018.11.20
request==2018.11.20
requests==2.20.1
requests-oauthlib==1.0.0
rope==0.11.0
rsa==4.0
ruamel.yaml==0.15.80
query-string==2018.11.20
request==2018.11.20
requests==2.20.1
requests-oauthlib==1.0.0
rope==0.11.0
rsa==4.0
six==1.11.0
typed-ast==1.1.0
upgrade-requirements==1.7.0
uritemplate==3.0.0
urllib3==1.24.1
upgrade-requirements==1.7.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
yapf==0.25.0
